+++ tests/testthat/test_Layout.R
@@ -33,20 +33,20 @@ test_that("LayoutFormat works as expected", {
 
 test_that("LayoutGlue works as expected", {
   # basic formatting works
-  lo <- LayoutGlue$new(fmt = "{level} [{timestamp}] msg")
-  expect_match(lo$format_event(tevent), "^200.*")
+  lo <- LayoutGlue$new(fmt = "{level} [{timestamp}] {msg}")
+  expect_match(lo$format_event(tevent), "^200.*foo bar$")
 
   # active bindings work
-  lo <- LayoutGlue$new(fmt = "{level_name} [{timestamp}] msg")
-  expect_match(lo$format_event(tevent), "^error.*")
+  lo <- LayoutGlue$new(fmt = "{level_name} [{timestamp}] {msg}")
+  expect_match(lo$format_event(tevent), "^error.*foo bar$")
 
   # functions work
-  lo <- LayoutGlue$new(fmt = "{toupper(level_name)} [{timestamp}] msg")
-  expect_match(lo$format_event(tevent), "^ERROR.*")
+  lo <- LayoutGlue$new(fmt = "{toupper(level_name)} [{timestamp}] {msg}")
+  expect_match(lo$format_event(tevent), "^ERROR.*foo bar$")
 
   # default format works
-  lo <- LayoutGlue$new(fmt = "{pad_right(colorize_levels(toupper(level_name)), 5)} [{timestamp}] msg")
-  expect_match(lo$format_event(tevent), "ERROR.*msg$")
+  lo <- LayoutGlue$new(fmt = "{pad_right(colorize_levels(toupper(level_name)), 5)} [{timestamp}] {msg}")
+  expect_match(lo$format_event(tevent), "ERROR.*foo bar$")
 
   if (crayon::has_color()){
     expect_true(crayon::has_style(lo$format_event(tevent)))